Selection Sort - 1.0
First commit, the program seems to be complete here, no
errors so far, even though the code is not well documented
it is well written and clear to understand. There should
be no problems even for relatively inexperienced people
to understand this logic since it is an intuitive way of
sorting. I do not intend to improve upon the code.